requirements.txt
google-generativeai>=0.4.0
numpy>=1.22.5
nltk>=3.8.1
scikit-learn>=1.2.2
gradio==5.29.0
gradio-calendar>=0.0.6
matplotlib>=3.7.0
pillow>=9.0.0
pydub>=0.25.1
speechrecognition>=3.10.0
wordcloud>=1.9.0
facenet-pytorch>=2.5.2
fer>=22.5.0
mtcnn>=0.1.1
typing-extensions>=4.4.0
tensorflow>=2.12.0
moviepy>=1.0.3 